Support device level power utilization metrics
The collector was only exporting the sense_energy_power_watts metric for the
'main' and 'solar' device names. This was only the top-level power consumption
and production. The API does appear to expose realtime power utilization data
for Sense-identified devices on the account, including the Always On and Other
aggregate buckets.

This change enables exporting more 'name' label values for individual device
level utilization on the sense_energy_power_watts metric.

Fixes #3.